Refactor crinit util package into multiple packages
